feat: JWT 기반 회원가입/로그인 구현 및 사용자 별 계약 분석 저장 로직 고도화

+++ backend/src/main/java/com/ailawyer/backend/dashboard/entity/UserEntity.java
@@ -14,7 +14,7 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Integer userId; // PK
+    private Long userId; // PK
     @Column(name = "password")
     private String password;
     @Column(name = "nickname")
